+++ 0x02-functions_nested_loops/main.h
@@ -3,5 +3,5 @@ void print_alphabet(void);
 void print_alphabet_x10(void);
 int _islower(int c);
 int _isalpha(int c);
-signed long int print_sign(signed long int n)
+signed int print_sign(signed int n)
 int _abs(int);
